rename Label files for consistency
