+++ src/components/Label/Label.mdx
@@ -1,5 +1,10 @@
-import { Preview, Story, ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs';
-import LobLabel from './LobLabel.vue';
+import {
+  Preview,
+  Story,
+  ArgsTable,
+  PRIMARY_STORY,
+} from "@storybook/addon-docs";
+import LobLabel from "./Label.vue";
 
 # Lob Label
 
@@ -18,12 +23,12 @@ If the field is required, you can pass in the `required` prop to show the asteri
 You can hide the label by setting the `srOnlyLabel` to true (it will still be read by a screen reader).
 
 ```html
-  <lob-label  
-    :label="label"
-    :labelFor="labelFor"
-    :required="required"
-    :srOnlyLabel="srOnlyLabel"
-  />
+<lob-label
+  :label="label"
+  :labelFor="labelFor"
+  :required="required"
+  :srOnlyLabel="srOnlyLabel"
+/>
 ```
 
 ## Props
