+++ src/components/Label/Label.vue
