+++ src/components/Label/__tests__/LobLabel.spec.js
